:recycle: Read num of blackboard finish
