+++ src/main/resources/application.properties
@@ -7,6 +7,6 @@ spring.jpa.show-sql=true
 spring.datasource.url=jdbc:mysql://localhost:3306/black_board
 spring.datasource.username=root
 spring.datasource.password=1234
-spring.jpa.hibernate.ddl-auto=create
+spring.jpa.hibernate.ddl-auto=update
 # Swagger2
 spring.mvc.pathmatch.matching-strategy=ant_path_matcher
